requirements.txt
typing
psycopg2==2.8.3
psycopg2-binary==2.8.3
PyMySQL==1.0.2
SQLAlchemy==1.3.8
gunicorn==20.0.4
fastapi==0.103.1
pydantic==2.5.2 
starlette==0.27.0
websockets==11.0
uvicorn==0.22.0
cloud-sql-python-connector[pymysql]
graphql-core==3.2.3
strawberry-graphql==0.196.0
smartystreets_python_sdk